# sycl : Reenabled mmvq path for the SYCL Nvidia Backend