@Alcpz 2024-07-08
@OuadiElfarouki @joeatodd @AidanBeltonS tagging for the discussion wrt the backend.

@airMeng @NeoZhangJianyu: What do you think of this? Is the change ok for the multi-device implementation?


I considered the alternative of re-adding a macro to "enable mmvq", but a dynamic check seems simpler as we already have plenty of Macros.